session_2/challenge/requirements.txt
g4f>=0.2.9.9
tqdm>=4.66.2
absl-py>=2.1.0